Add files via upload
In this code snippet....
The index.php contains basic html code. Three input boxes are used for Name, Email Address, and Password required for the first step of registration.
The code.php establishes connection with database and stores data of first step of registration in database.
The login.php contains html code for login purpose.
The logcode.php checks if the user is already registered and valid for login. If user is valid it allows access for further steps.
The finalreg.php is for final part of registration and contains html code.
The finalcode.php establishes connection with database and store second required data to complete the registration.
Through update.php and updatecode.php we can update/ make change in our data if we filled wrong data by mistake.
The view.php helps us to see all registered data/ all registrations.